Refactor use cases into separate files with backward compatibility
